refactor(hooks): test routing state through its own module
The routing state cells were extracted into routing-state.ts by the hooks
cycle removal, but the tests were never moved. event-router.ts carried a
re-export whose stated purpose was to keep itself the single import
surface the test suite addresses -- an export that existed for tests
rather than for production.

Repoints 14 test files at routing-state.ts for the four symbols it owns:
currentEpoch, getRoutingBucket, PendingSessionStartContext and
RoutingEntry. Symbols event-router.ts genuinely owns stay imported from
it. Four orchestrator tests reach it through a dynamic import and now
take getRoutingBucket from routing-state.ts alongside their existing
event-router destructure.

One production consumer also relied on the re-export: spawn-helpers.ts
imported RoutingEntry from event-router.ts and now imports it from
routing-state.ts directly. The re-export is deleted, and getRoutingBucket
is no longer imported by event-router.ts at all.

One test moved rather than being repointed. Its subject was currentEpoch
itself, so it belongs beside the module, in the new
tests/bridges/hooks/routing-state.test.ts. Tests that merely use these
symbols while asserting event-router behaviour stayed put -- those
exercise event-router's public interface and the symbols are helpers.

CONVENTIONS.md records the principle in order: test the public interface;
treat pressure to reach inside as signal a module wants to exist; extract
only when the inner code is semantically meaningful, otherwise go back to
the public interface. Export-for-test is named as the anti-pattern that
relieves the pressure without doing the design work. Dependency injection
is presented as the principle applied, not an exception.

Files FLOW-09 for the remaining ~84 exports and ~39 seams, framed as
decisions with two legitimate outcomes -- promote to a module, or delete
the export and test through the public interface -- not as 84 pending
extractions.
