+++ tests/bridges/hooks/settle.test.ts
@@ -15,13 +15,15 @@ import {
   _setExecutorForTest,
 } from "../../../extensions/pi-claude-marketplace/bridges/hooks/dispatch.ts";
 import {
-  currentEpoch,
   _bumpEpochForTest,
   _resetForTest,
   _setRoutingBucketForTest,
-  type RoutingEntry,
 } from "../../../extensions/pi-claude-marketplace/bridges/hooks/event-router.ts";
 import { MATCH_ALL_IF } from "../../../extensions/pi-claude-marketplace/bridges/hooks/if-field/index.ts";
+import {
+  currentEpoch,
+  type RoutingEntry,
+} from "../../../extensions/pi-claude-marketplace/bridges/hooks/routing-state.ts";
 import {
   agentEndCacheHandler,
   inputResetHandlerFor,
